requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
httpx>=0.25.0
pillow>=10.0.0
pdf2image>=1.16.0
PyPDF2>=3.0.0
pydantic>=2.8.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0
aiofiles>=23.2.0
python-magic>=0.4.27
asyncio-throttle>=1.0.0
